+++ src/types/config.ts
@@ -1,7 +1,7 @@
 import type { DateTimeUnit, DurationLike } from "luxon";
 import { PERIOD_TYPES, type PeriodType } from "../constants";
 import type { PeriodChildren, PeriodLinks } from "./period";
-import type { DirectorySettings, NamingSettings } from "./schemas";
+import type { DirectorySettings, NamingSettings, TemplaterSettings } from "./schemas";
 
 export const ORDERED_PERIOD_TYPES: PeriodType[] = [
 	PERIOD_TYPES.YEARLY,
@@ -19,6 +19,7 @@ export interface PeriodConfig {
 	duration: DurationLike;
 	folderKey: keyof DirectorySettings;
 	formatKey: keyof NamingSettings;
+	templateKey: keyof Omit<TemplaterSettings, "enabled">;
 	linkKey: LinkKey | null;
 	childrenKey: ChildrenKey | null;
 	parent: PeriodType | null;
@@ -31,6 +32,7 @@ export const PERIOD_CONFIG: Record<PeriodType, PeriodConfig> = {
 		duration: { days: 1 },
 		folderKey: "dailyFolder",
 		formatKey: "dailyFormat",
+		templateKey: "dailyTemplate",
 		linkKey: null,
 		childrenKey: "days",
 		parent: "weekly",
@@ -41,6 +43,7 @@ export const PERIOD_CONFIG: Record<PeriodType, PeriodConfig> = {
 		duration: { weeks: 1 },
 		folderKey: "weeklyFolder",
 		formatKey: "weeklyFormat",
+		templateKey: "weeklyTemplate",
 		linkKey: "week",
 		childrenKey: "weeks",
 		parent: "monthly",
@@ -51,6 +54,7 @@ export const PERIOD_CONFIG: Record<PeriodType, PeriodConfig> = {
 		duration: { months: 1 },
 		folderKey: "monthlyFolder",
 		formatKey: "monthlyFormat",
+		templateKey: "monthlyTemplate",
 		linkKey: "month",
 		childrenKey: "months",
 		parent: "quarterly",
@@ -61,6 +65,7 @@ export const PERIOD_CONFIG: Record<PeriodType, PeriodConfig> = {
 		duration: { quarters: 1 },
 		folderKey: "quarterlyFolder",
 		formatKey: "quarterlyFormat",
+		templateKey: "quarterlyTemplate",
 		linkKey: "quarter",
 		childrenKey: "quarters",
 		parent: "yearly",
@@ -71,6 +76,7 @@ export const PERIOD_CONFIG: Record<PeriodType, PeriodConfig> = {
 		duration: { years: 1 },
 		folderKey: "yearlyFolder",
 		formatKey: "yearlyFormat",
+		templateKey: "yearlyTemplate",
 		linkKey: "year",
 		childrenKey: null,
 		parent: null,